Fix deletion and addition bug by refactoring
Design of code has been rewritten. The flow of functions is as follows:
First, masterArray is updated. Then, the localStorage is changed ONLY
if masterArray is changed. Lastly, DOM is changed but ONLY if
masterArray is changed.

Essentially, the DOM does not communicate with the localStorage and vice
versa. masterArray is what changes both, INDEPENDENTLY.